pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-hospital-scheduler"
version = "2.0.0"
description = "Hospital appointment scheduling environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "pyyaml>=6.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
demo = [
    "groq>=0.4.0",
    "python-dotenv>=1.0.0",
    "gradio>=4.0.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
server = "hospital_scheduler.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["hospital_scheduler", "hospital_scheduler.server"]
package-dir = { "hospital_scheduler" = ".", "hospital_scheduler.server" = "server" }
